refactoring login and register to auth controller
